Fix option merging when provider is a string
This is the result of @giddie discovery and work in #18

In some cases, the incoming provider name may be a string (like when using `run_request` directly)
instead of an atom, as typically seen in the docs. In those cases,
this fails to get the provider options that may be in the config.

This fixes the lookup to be indifferent by forcing both keys to be strings
when comparing